requirements.txt
aws-cdk-lib~=2.63.0
constructs>=10.1.0,<11.0.0
aws-cdk.aws-lambda-python-alpha
-r lambda/requirements.txt
